Turn tool debugging always on.
Rationale:
==========
The TOOL is always in a state of development and debugging.

Most people on the team are unaware you can do this, so they get useless stack traces.

The TOOL (especially the Java side) performs very little intensive calculation, and so this should have little or no performance impact.